100 - move data sources to separate folder
