+++ tests/tel2puml/otel_to_pv/test_ingest_data.py
@@ -19,7 +19,7 @@ from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
 from tel2puml.otel_to_pv.data_holders.otel_data_holder import (
     SQLDataHolder,
 )
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_source import (
+from tel2puml.otel_to_pv.data_sources.otel_data_source import (
     JSONDataSource,
 )
