+++ tests/tel2puml/otel_to_pv/test_otel_data_source.py
@@ -10,7 +10,7 @@ from pytest import FixtureRequest
 from pathlib import Path
 from unittest.mock import mock_open, patch
 
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_source import (
+from tel2puml.otel_to_pv.data_sources.otel_data_source import (
     JSONDataSource,
 )
 from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
